Remove duplicate download_experts.py
