+++ experiments/download_experts.py
@@ -1,13 +1,31 @@
-import urllib
+import urllib.request
 import zipfile
 
-EXPERTS_URL = ("http://www.dl.dropboxusercontent.com/"
-               "s/hww9n76d242433c/stable_baselines_expert_models.zip")
+experts_url = ("https://www.dl.dropboxusercontent.com/"
+               "s/u3uuyd8nxs7obwp/expert_policies.zip")
 
-def main():
-  zip_path, _ = urllib.urlretrieve(EXPERTS_URL, dest_zip_path)
-  with zipfile.ZipFile(zip_path,"r") as zip_ref:
+experts_dest = "expert_models/"
+
+
+def main(verbose=True):
+  try:
+    zip_path, _ = urllib.request.urlretrieve(experts_url)
+  except urllib.error.URLError as e:
+    if "SSL" in str(e):
+      print(
+        "Suggestion: If on macOS, run "
+        r"`/Applications/Python\ 3.6/Install\ Certificates.command`"
+        "(https://stackoverflow.com/a/13531310/1091722)")
+    raise e
+
+  if verbose:
+    print("Downloaded expert_policies.zip")
+
+  with zipfile.ZipFile(zip_path, "r") as zip_ref:
     zip_ref.extractall("expert_models")
+    if verbose:
+      print(f"Extracted the following experts to {experts_dest}:")
+      zip_ref.printdir()
 
 
 if __name__ == "__main__":
